Rename `utils/utils` -> `utils` and `map/map` -> `map`
