+++ src/items/itemGenerator.ts
@@ -8,14 +8,14 @@ import Pickable, {
   SelectorType,
   TargetSelector,
   Wearable,
-  WearableType
+  WearableType,
 } from "@/items/pickable";
 import Armor, { ArmorType } from "@/rpg/armor";
 import Weapon, { DamageType } from "@/rpg/weapon";
 import Actor from "@/units/actor";
 import { ConfusedMonsterAi } from "@/units/ai";
+import { ensure } from "@/utils";
 import { Colors } from "@/utils/colors";
-import { ensure } from "@/utils/utils";
 
 interface ArmorInterface {
   name: string;
